Add content verification token
This is how we’ll verify a user clicked on a valid email verification
link.
